split yjs store and react-flow store
